dataset: Extract PackingList into its own module
