+++ src/dataset.rs
@@ -16,3 +16,6 @@ pub trait Dataset<LogicalRecord> {
 
 mod indexed;
 pub use indexed::*;
+
+mod packing_list;
+pub use packing_list::*;
