+++ src/schema.rs
@@ -21,6 +21,18 @@ impl<S: AsRef<str>> core::convert::From<S> for OldSchema {
 	}
 }
 
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+#[non_exhaustive]
+pub enum Schema {
+	Census2010(census2010::Schema),
+}
+
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+#[non_exhaustive]
+pub enum Table {
+	Census2010(census2010::Table),
+}
+
 #[cfg(test)]
 mod tests {
 	use super::OldSchema;
